K8s CI Runner [1st phase]

Changes for hybrid docker/k8s runner support.

`HYBRID_FACTOR_{SF,ZH,CH}` is expected that somehow defines the ratio of some jobs that we want to distribute to k8s runners. If the child pipeline has pinned jobs assigned to SF and `HYBRID_FACTOR_SF` is non-negative, then we distribute some jobs to k8s runners. If `HYBRID_FACTOR_SF` is 50, there's a 50% chance that all the jobs from `cargo-build` (exception are darwin jobs), `cargo-test` and `build-determinism` stage will be distributed to k8s runners. The same goes for `HYBRID_FACTOR_ZH` and `HYBRID_FACTOR_CH`.

We install `nix` in `before_script` section if we notice that it's missing - when new node is added to k8s node pool. We should make changes in cluster-api to do this when we add the node and latter remove this logic.

Covers 1st class of CI jobs (or most of them) in `cargo-pipeline` described in [IDX-2404](https://dfinity.atlassian.net/browse/IDX-2404). 

See merge request dfinity-lab/public/ic!6516

+++ gitlab-ci/src/artifacts/openssl-sign.sh
@@ -21,6 +21,11 @@ rm -f SHA256SUMS sign-input.txt sign.sig sign.sig.bin
     sha256sum -b * | tee SHA256SUMS
 )
 
+if [ "${CI_COMMIT_REF_PROTECTED:-}" != "true" ] && [ ! -e /openssl/private.pem ]; then
+    echo "Not doing anything as /openssl/private.pem doesn't exist. [NOT on protected branch]"
+    exit 0
+fi
+
 cp SHA256SUMS sign-input.txt
 VERSION="${VERSION:-$(git rev-parse --verify HEAD)}"
 echo "$VERSION" >>sign-input.txt
